+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+    "python.linting.enabled": true,
+    "python.pythonPath": "C:\\Users\\Redouane\\AppData\\Local\\Programs\\Python\\Python39\\python.exe"
+}
