+++ chess_py_development-main/.vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "python.pythonPath": "C:\\Users\\Redouane\\AppData\\Local\\Programs\\Python\\Python39\\python.exe"
+}
