+++ chess_py_development-main/assets/black_queen.png
