+++ chess_py_development-main/assets/black_rook.png
