+++ chess_py_development-main/assets/white_knight.png
