+++ chess_py_development-main/assets/white_pawn.png
