+++ chess_py_development-main/assets/white_rook.png
