+++ chess_py_development-main/components/__pycache__/pieces_cls.cpython-38.pyc
