+++ chess_py_development-main/components/__pycache__/pieces_test_cls.cpython-38.pyc
