+++ chess_py_development-main/test.py
@@ -1,0 +1,4 @@
+list = [1, 2, 3, 'dbas', 5]
+
+for 2 in list:
+    print('hi')
