+++ setup.config
@@ -1,0 +1,3 @@
+[pycodestyle]
+max_line_length = 120
+ignore = E501
