--- test.py
@@ -1,4 +1,0 @@
-list = [1, 2, 3, 'dbas', 5]
-
-for 2 in list:
-    print('hi')
